% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpdf_document_base.R
\name{hpdf_document_base}
\alias{hpdf_document_base}
\title{A basic PDF document output format}
\usage{
hpdf_document_base(toc = FALSE, toc_depth = 3, section_divs = TRUE,
  fig_width = 5, fig_height = 4, fig_retina = 8, fig_caption = TRUE,
  dev = "png", dpi = 96, df_print = NULL, notes = c("endnotes",
  "footnotes"), attach_code = FALSE, smart = TRUE, highlight = "default",
  math_engine = "webtex_svg", template = NULL, html_format = c("html4",
  "html5"), engine = c("weasyprint", "prince"), engine_opts = NULL,
  extra_dependencies = NULL, css = NULL, includes = NULL,
  keep_md = FALSE, keep_html = FALSE, self_contained = TRUE,
  lib_dir = NULL, md_extensions = NULL, pandoc_args = NULL, ...)
}
\arguments{
\item{toc}{\code{TRUE} to include a table of contents in the output}

\item{toc_depth}{Depth of headers to include in table of contents}

\item{section_divs}{Wrap sections in <div> tags (or <section> tags in HTML5),
and attach identifiers to the enclosing <div> (or <section>) rather than the
header itself.}

\item{fig_width}{Default width (in inches) for figures}

\item{fig_height}{Default width (in inches) for figures}

\item{fig_retina}{Setting this option to a ratio (for example, 2) will
change the \code{dpi} parameter to \code{dpi * fig.retina}, and \code{fig_width} to
\code{fig_width * dpi / fig_retina} internally; for example, the physical size
of an image is doubled and its display size is halved when \code{fig_retina = 2}.}

\item{fig_caption}{\code{TRUE} to render figures with captions}

\item{dev}{Graphics device to use for figure output (defaults to png)}

\item{dpi}{The DPI (dots per inch) for bitmap devices.}

\item{df_print}{Method to be used for printing data frames. Valid values
include \code{"default"}, \code{"kable"} and \code{"tibble"}. The \code{"default"} method uses
\code{print.data.frame}. The \code{"kable"} method uses the \code{\link[knitr:kable]{knitr::kable()}}
function. The \code{"tibble"} method uses the \link[tibble:tibble-package]{tibble::tibble-package} to print a
summary of the data frame. In addition to the named methods you can also
pass an arbitrary function to be used for printing data frames. You can
disable the \code{df_print} behavior entirely by setting the option
\code{rmarkdown.df_print} to \code{FALSE}.}

\item{notes}{Render notes as endnotes or footnotes. Footnotes are not
yet supported by WeasyPrint and are only well rendered with Prince.}

\item{attach_code}{Add the \code{Rmd} source code as an attachment to the \code{PDF}
document.}

\item{smart}{Produce typographically correct output, converting straight
quotes to curly quotes, --- to em-dashes, -- to en-dashes, and ... to
ellipses.}

\item{highlight}{Syntax highlighting style. Supported styles include
"default", "tango", "pygments", "kate", "monochrome", "espresso", "zenburn",
"haddock", and "textmate". Pass \code{NULL} to prevent syntax highlighting.}

\item{math_engine}{Method to be used to render \code{TeX} math. Valid values
include \code{"unicode"}, \code{"mathjax"}, \code{"mathml"}, \code{"webtex_svg"}, \code{"webtex_png"}
and \code{"katex"}. See the
\href{https://pandoc.org/MANUAL.html#math-rendering-in-html}{pandoc manual}
for details.}

\item{template}{\code{Pandoc} template to use for rendering. Pass \code{NULL} to use
\code{pandoc}'s built-in template; pass a path to use a custom template that
you've created. See the documentation on
\href{http://pandoc.org/MANUAL.html#templates}{pandoc online documentation}
for details on creating custom templates.}

\item{html_format}{HTML format of the intermediate HTML file.}

\item{engine}{A \code{PDF} engine.}

\item{engine_opts}{A character vector of command-line arguments to be passed to \code{engine}.}

\item{extra_dependencies}{Additional function arguments to pass to the
base R Markdown HTML output formatter \code{\link{html_document_base}}}

\item{css}{One or more css files to include}

\item{includes}{Named list of additional content to include within the
document (typically created using the \code{\link{includes}} function).}

\item{keep_md}{Keep the markdown file generated by knitting.}

\item{keep_html, self_contained}{Keep intermediate \code{HTML} file. Use
\code{self_contained} to indicate if external dependencies are embedded in
\code{HTML} file.}

\item{lib_dir}{Directory to copy dependent HTML libraries (e.g. jquery,
bootstrap, etc.) into. By default this will be the name of the document with
\code{_files} appended to it.}

\item{md_extensions}{Markdown extensions to be added or removed from the
default definition or R Markdown. See the \code{\link{rmarkdown_format}} for
additional details.}

\item{pandoc_args}{Additional command line options to pass to pandoc}

\item{...}{Additional arguments passed to \code{\link[rmarkdown:html_document_base]{rmarkdown::html_document_base()}}.}
}
\description{
A basic PDF document output format
}
