% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html2pdf.R
\name{html2pdf}
\alias{html2pdf}
\title{Add conversion to pdf}
\usage{
html2pdf(base_format)
}
\arguments{
\item{base_format}{Any \code{HTML} format.}
}
\value{
R Markdown output format to pass to \code{rmarkdown::render}.
}
\description{
Add conversion to pdf
}
