% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pdf.R
\name{make_pdf}
\alias{make_pdf}
\title{Generate a pdf from an html file}
\usage{
make_pdf(file, engine = c("weasyprint", "prince"), engine_opts = NULL,
  pdf_file = gsub("html?$", "pdf", file))
}
\arguments{
\item{file}{An \code{HTML} file path.}

\item{engine}{A \code{PDF} engine.}

\item{engine_opts}{A character vector of command-line arguments to be passed to \code{engine}.}

\item{pdf_file}{Path to the \code{PDF} output file. By default, it is under the
same directory as the \code{input} file and also has the same base name.}
}
\value{
A character string of the path of the \code{PDF} output file (i.e., the
value of the \code{pdf_file} argument).
}
\description{
This function is a helper to use \href{https://weasyprint.org/}{WeasyPrint} or
\href{https://www.princexml.com/}{Prince} in order to convert an \code{HTML} file to
\code{PDF}.
}
\details{
The \href{https://weasyprint.readthedocs.io/en/latest/api.html#cmdoption-p}{presentational hints option}
is always enabled for \code{Weasyprint}.
The \href{https://www.princexml.com/doc-refs/#cl-javascript}{javascript option}
is always enabled for \code{Prince}.
}
