% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_document_base.R
\name{pdf_document_base}
\alias{pdf_document_base}
\title{A simple pdf generator}
\usage{
pdf_document_base(fig_width = 5, fig_height = 4, fig_caption = TRUE,
  smart = TRUE, keep_md = FALSE, md_extensions = NULL,
  pdf_engine = "weasyprint", verbose = FALSE)
}
\arguments{
\item{fig_width}{Default width (in inches) for figures}

\item{fig_height}{Default width (in inches) for figures}

\item{fig_caption}{\code{TRUE} to render figures with captions}

\item{smart}{Produce typographically correct output, converting straight
quotes to curly quotes, --- to em-dashes, -- to en-dashes, and ... to
ellipses.}

\item{keep_md}{Keep the markdown file generated by knitting.}

\item{md_extensions}{Markdown extensions to be added or removed from the
default definition or R Markdown. See the \code{\link{rmarkdown_format}} for
additional details.}

\item{pdf_engine}{PDF engine passed to pandoc}

\item{verbose}{logical indicating if verbose option passed to pandoc}
}
\value{
R Markdown output format to pass to \code{rmarkdown::render}.
}
\description{
A simple pdf generator
}
