% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpdf_book.R
\name{wpdf_document2}
\alias{wpdf_document2}
\title{Output format for producing a bookdown pdf}
\usage{
wpdf_document2(..., engine = c("weasyprint", "prince"), engine_opts = NULL,
  number_sections = TRUE, pandoc_args = NULL, notes = c("endnotes",
  "footnotes"), base_format = rmarkdown::html_document)
}
\arguments{
\item{...}{Arguments to be passed to a
specific output format function. For a function \code{foo2()}, its
arguments are passed to \code{foo()}, e.g. \code{...} of
\code{html_document2()} are passed to \code{rmarkdown::html_document()}.}

\item{engine}{A \code{PDF} engine.}

\item{engine_opts}{A character vector of command-line arguments to be passed to \code{engine}.}

\item{number_sections}{Whether to number section headers: if \code{TRUE},
figure/table numbers will be of the form \code{X.i}, where \code{X} is the
current first-level section number, and \code{i} is an incremental number
(the i-th figure/table); if \code{FALSE}, figures/tables will be numbered
sequentially in the document from 1, 2, ..., and you cannot cross-reference
section headers in this case.}

\item{pandoc_args}{Arguments to be passed to a
specific output format function. For a function \code{foo2()}, its
arguments are passed to \code{foo()}, e.g. \code{...} of
\code{html_document2()} are passed to \code{rmarkdown::html_document()}.}

\item{notes}{Render notes as endnotes or footnotes. Footnotes are not
yet supported by WeasyPrint and are only well rendered with Prince.}

\item{base_format}{An output format function to be used as the base format.}
}
\value{
An R Markdown output format object to be passed to
\code{\link[rmarkdown:render]{rmarkdown::render()}}.
}
\description{
Output format for producing a bookdown pdf
}
